Start and stop collecting measurements behavior
Also a refactor on the files disposition.
