+++ lib/connection/ble/bluetooth_widgets.dart
