+++ lib/pages/devices_connection_page.dart
@@ -3,7 +3,8 @@ import 'dart:async';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
-import 'bluetooth_device.dart';
+import 'package:lsa_gloves/connection/ble/bluetooth_specification.dart';
+import '../connection/ble/bluetooth_device.dart';
 
 class GloveConnectionPage extends StatelessWidget {
   @override
@@ -57,84 +58,94 @@ class FindDevicesScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-        appBar: AppBar(
+      appBar: AppBar(
         title: Text('Find Devices'),
-    ),
-    body: RefreshIndicator(
-      onRefresh: () =>
-          FlutterBlue.instance.startScan(timeout: Duration(seconds: 4)),
-      child: SingleChildScrollView(
-        child: Column(
-          children: <Widget>[
-            StreamBuilder<List<BluetoothDevice>>(
-              stream: Stream.periodic(Duration(seconds: 2))
-                  .asyncMap((_) => FlutterBlue.instance.connectedDevices),
-              initialData: [],
-              builder: (c, snapshot) => Column(
-                children: snapshot.data!.map((d) => ListTile(
-                  title: Text(d.name),
-                  subtitle: Text(d.id.toString()),
-                  trailing: StreamBuilder<BluetoothDeviceState>(
-                    stream: d.state,
-                    initialData: BluetoothDeviceState.disconnected,
-                    builder: (c, snapshot) {
-                      if (snapshot.data ==
-                          BluetoothDeviceState.connected) {
-                        return RaisedButton(
-                          child: Text('OPEN'),
-                          onPressed: () => Navigator.of(context).push(
-                              MaterialPageRoute(
-                                  builder: (context) =>
-                                      DeviceScreen(device: d))),
-                        );
-                      }
-                      return Text(snapshot.data.toString());
-                    },
-                  ),
-                ))
-                    .toList(),
+      ),
+      body: RefreshIndicator(
+        onRefresh: () =>
+            FlutterBlue.instance.startScan(timeout: Duration(seconds: 4)),
+        child: SingleChildScrollView(
+          child: Column(
+            children: <Widget>[
+              StreamBuilder<List<BluetoothDevice>>(
+                stream: Stream.periodic(Duration(seconds: 2))
+                    .asyncMap((_) => FlutterBlue.instance.connectedDevices),
+                initialData: [],
+                builder: (c, snapshot) => Column(
+                  children: snapshot.data!
+                      .map((d) => ListTile(
+                            title: Text(d.name),
+                            subtitle: Text(d.id.toString()),
+                            trailing: StreamBuilder<BluetoothDeviceState>(
+                              stream: d.state,
+                              initialData: BluetoothDeviceState.disconnected,
+                              builder: (c, snapshot) {
+                                if (snapshot.data ==
+                                    BluetoothDeviceState.connected) {
+                                  return RaisedButton(
+                                    child: Text('OPEN'),
+                                    onPressed: () => Navigator.of(context).push(
+                                        MaterialPageRoute(
+                                            builder: (context) =>
+                                                DeviceScreen(device: d))),
+                                  );
+                                }
+                                return Text(snapshot.data.toString());
+                              },
+                            ),
+                          ))
+                      .toList(),
+                ),
               ),
-            ),
-            StreamBuilder<List<ScanResult>>(
-              stream: FlutterBlue.instance.scanResults,
-              initialData: [],
-              builder: (c, snapshot) => Column(
-                children: snapshot.data!
-                    .map(
-                      (r) => ScanResultTile(
-                    result: r,
-                    onTap: () => Navigator.of(context)
-                        .push(MaterialPageRoute(builder: (context) {
-                      r.device.connect();
-                      return DeviceScreen(device: r.device);
-                    })),
-                  ),
-                )
-                    .toList(),
+              StreamBuilder<List<ScanResult>>(
+                stream: FlutterBlue.instance.scanResults,
+                initialData: [],
+                builder: (c, snapshot) => Column(
+                  children: filterGloveDevices(snapshot)
+                      .map(
+                        (r) => ScanResultTile(
+                          result: r,
+                          onTap: () => Navigator.of(context)
+                              .push(MaterialPageRoute(builder: (context) {
+                            r.device.connect();
+                            return DeviceScreen(device: r.device);
+                          })),
+                        ),
+                      )
+                      .toList(),
+                ),
               ),
-            ),
-          ],
+            ],
+          ),
         ),
       ),
-    ),
-    floatingActionButton: StreamBuilder<bool>(
-    stream: FlutterBlue.instance.isScanning,
-      initialData: false,
-      builder: (c, snapshot) {
-        if (snapshot.data!) {
-          return FloatingActionButton(
-            child: Icon(Icons.stop),
-            onPressed: () => FlutterBlue.instance.stopScan(),
-            backgroundColor: Colors.red,
-          );
-        } else {
-          return FloatingActionButton(
-              child: Icon(Icons.search),
-              onPressed: () => FlutterBlue.instance.startScan(timeout: Duration(seconds: 4)));
-        }
-      },
-    ),
+      floatingActionButton: StreamBuilder<bool>(
+        stream: FlutterBlue.instance.isScanning,
+        initialData: false,
+        builder: (c, snapshot) {
+          if (snapshot.data!) {
+            return FloatingActionButton(
+              child: Icon(Icons.stop),
+              onPressed: () => FlutterBlue.instance.stopScan(),
+              backgroundColor: Colors.red,
+            );
+          } else {
+            return FloatingActionButton(
+                child: Icon(Icons.search),
+                onPressed: () => FlutterBlue.instance
+                    .startScan(timeout: Duration(seconds: 4)));
+          }
+        },
+      ),
     );
+  }
+
+  List<ScanResult> filterGloveDevices(
+      AsyncSnapshot<List<ScanResult>> snapshot) {
+    return snapshot.data!
+        .where((element) =>
+            element.device.name == BluetoothSpecification.deviceName)
+        .toList();
   }
 }
 
@@ -249,5 +260,3 @@ class ScanResultTile extends StatelessWidget {
     );
   }
 }
-
-
